+++ examples_dart/bin/omemo_client.dart
@@ -30,7 +30,7 @@ void main(List<String> args) async {
     TestingReconnectionPolicy(),
     AlwaysConnectedConnectivityManager(),
     ClientToServerNegotiator(),
-    ExampleTCPSocketWrapper(parser.srvRecord),
+    ExampleTCPSocketWrapper(parser.srvRecord, true),
   )..connectionSettings = parser.connectionSettings;
 
   // Generate OMEMO data