+++ packages/moxxmpp/lib/src/awaiter.dart
@@ -1,35 +1,13 @@
 import 'dart:async';
-import 'package:meta/meta.dart';
-import 'package:moxxmpp/src/jid.dart';
 import 'package:moxxmpp/src/stringxml.dart';
 import 'package:synchronized/synchronized.dart';
 
-/// A surrogate key for awaiting stanzas.
-@immutable
-class _StanzaSurrogateKey {
-  const _StanzaSurrogateKey(this.sentTo, this.id, this.tag);
+/// (JID we sent a stanza to, the id of the sent stanza, the tag of the sent stanza).
+// ignore: avoid_private_typedef_functions
+typedef _StanzaCompositeKey = (String?, String, String);
 
-  /// The JID the original stanza was sent to. We expect the result to come from the
-  /// same JID.
-  final String sentTo;
-
-  /// The ID of the original stanza. We expect the result to have the same ID.
-  final String id;
-
-  /// The tag name of the stanza.
-  final String tag;
-
-  @override
-  int get hashCode => sentTo.hashCode ^ id.hashCode ^ tag.hashCode;
-
-  @override
-  bool operator ==(Object other) {
-    return other is _StanzaSurrogateKey &&
-        other.sentTo == sentTo &&
-        other.id == id &&
-        other.tag == tag;
-  }
-}
+/// Callback function that returns the bare JID of the connection as a String.
+typedef GetBareJidCallback = String Function();
 
 /// This class handles the await semantics for stanzas. Stanzas are given a "unique"
 /// key equal to the tuple (to, id, tag) with which their response is identified.
@@ -40,8 +18,12 @@ class _StanzaSurrogateKey {
 ///
 /// This class also handles some "edge cases" of RFC 6120, like an empty "from" attribute.
 class StanzaAwaiter {
+  StanzaAwaiter(this._bareJidCallback);
+
+  final GetBareJidCallback _bareJidCallback;
+
   /// The pending stanzas, identified by their surrogate key.
-  final Map<_StanzaSurrogateKey, Completer<XMLNode>> _pending = {};
+  final Map<_StanzaCompositeKey, Completer<XMLNode>> _pending = {};
 
   /// The critical section for accessing [StanzaAwaiter._pending].
   final Lock _lock = Lock();
@@ -52,30 +34,33 @@ class StanzaAwaiter {
   /// [tag] is the stanza's tag name.
   ///
   /// Returns a future that might resolve to the response to the stanza.
-  Future<Future<XMLNode>> addPending(String to, String id, String tag) async {
+  Future<Future<XMLNode>> addPending(String? to, String id, String tag) async {
+    // Check if we want to send a stanza to our bare JID and replace it with null.
+    final processedTo = to != null && to == _bareJidCallback() ? null : to;
+
     final completer = await _lock.synchronized(() {
       final completer = Completer<XMLNode>();
-      _pending[_StanzaSurrogateKey(to, id, tag)] = completer;
+      _pending[(processedTo, id, tag)] = completer;
       return completer;
     });
 
     return completer.future;
   }
 
-  /// Checks if the stanza [stanza] is being awaited. [bareJid] is the bare JID of
-  /// the connection.
+  /// Checks if the stanza [stanza] is being awaited.
   /// If [stanza] is awaited, resolves the future and returns true. If not, returns
   /// false.
-  Future<bool> onData(XMLNode stanza, JID bareJid) async {
-    assert(bareJid.isBare(), 'bareJid must be bare');
-
+  Future<bool> onData(XMLNode stanza) async {
     final id = stanza.attributes['id'] as String?;
     if (id == null) return false;
 
-    final key = _StanzaSurrogateKey(
-      // Section 8.1.2.1 § 3 of RFC 6120 says that an empty "from" indicates that the
-      // attribute is implicitly from our own bare JID.
-      stanza.attributes['from'] as String? ?? bareJid.toString(),
+    // Check if we want to send a stanza to our bare JID and replace it with null.
+    final from = stanza.attributes['from'] as String?;
+    final processedFrom =
+        from != null && from == _bareJidCallback() ? null : from;
+
+    final key = (
+      processedFrom,
       id,
       stanza.tag,
     );
@@ -90,5 +75,20 @@ class StanzaAwaiter {
 
       return false;
     });
+  }
+
+  /// Checks if [stanza] represents a stanza that is awaited. Returns true, if [stanza]
+  /// is awaited. False, if not.
+  Future<bool> isAwaited(XMLNode stanza) async {
+    final id = stanza.attributes['id'] as String?;
+    if (id == null) return false;
+
+    final key = (
+      stanza.attributes['from'] as String?,
+      id,
+      stanza.tag,
+    );
+
+    return _lock.synchronized(() => _pending.containsKey(key));
   }
 }