+++ packages/moxxmpp/lib/src/xeps/xep_0045/status_codes.dart
@@ -1,0 +1,2 @@
+const selfPresenceStatus = '110';
+const nicknameChangedStatus = '303';