+++ packages/moxxmpp/test/awaiter_test.dart
@@ -2,11 +2,12 @@ import 'package:moxxmpp/moxxmpp.dart';
 import 'package:moxxmpp/src/awaiter.dart';
 import 'package:test/test.dart';
 
-void main() {
-  const bareJid = JID('moxxmpp', 'server3.example', '');
+const bareJid = 'user4@example.org';
+String getBareJidCallback() => bareJid;
 
+void main() {
   test('Test awaiting an awaited stanza with a from attribute', () async {
-    final awaiter = StanzaAwaiter();
+    final awaiter = StanzaAwaiter(getBareJidCallback);
 
     // "Send" a stanza
     final future = await awaiter.addPending(
@@ -20,14 +21,12 @@ void main() {
       XMLNode.fromString(
         '<iq from="user3@server.example" id="abc123" type="result" />',
       ),
-      bareJid,
     );
     expect(result1, false);
     final result2 = await awaiter.onData(
       XMLNode.fromString(
         '<iq from="user1@server.example" id="lol" type="result" />',
       ),
-      bareJid,
     );
     expect(result2, false);
 
@@ -37,22 +36,20 @@ void main() {
     );
     final result3 = await awaiter.onData(
       stanza,
-      bareJid,
     );
     expect(result3, true);
     expect(await future, stanza);
   });
 
   test('Test awaiting an awaited stanza without a from attribute', () async {
-    final awaiter = StanzaAwaiter();
+    final awaiter = StanzaAwaiter(getBareJidCallback);
 
     // "Send" a stanza
-    final future = await awaiter.addPending(bareJid.toString(), 'abc123', 'iq');
+    final future = await awaiter.addPending(null, 'abc123', 'iq');
 
     // Receive the wrong answer
     final result1 = await awaiter.onData(
       XMLNode.fromString('<iq id="lol" type="result" />'),
-      bareJid,
     );
     expect(result1, false);
 
@@ -60,23 +57,21 @@ void main() {
     final stanza = XMLNode.fromString('<iq id="abc123" type="result" />');
     final result2 = await awaiter.onData(
       stanza,
-      bareJid,
     );
     expect(result2, true);
     expect(await future, stanza);
   });
 
   test('Test awaiting a stanza that was already awaited', () async {
-    final awaiter = StanzaAwaiter();
+    final awaiter = StanzaAwaiter(getBareJidCallback);
 
     // "Send" a stanza
-    final future = await awaiter.addPending(bareJid.toString(), 'abc123', 'iq');
+    final future = await awaiter.addPending(null, 'abc123', 'iq');
 
     // Receive the correct answer
     final stanza = XMLNode.fromString('<iq id="abc123" type="result" />');
     final result1 = await awaiter.onData(
       stanza,
-      bareJid,
     );
     expect(result1, true);
     expect(await future, stanza);
@@ -84,31 +79,55 @@ void main() {
     // Receive it again
     final result2 = await awaiter.onData(
       stanza,
-      bareJid,
     );
     expect(result2, false);
   });
 
   test('Test ignoring a stanza that has the wrong tag', () async {
-    final awaiter = StanzaAwaiter();
+    final awaiter = StanzaAwaiter(getBareJidCallback);
 
     // "Send" a stanza
-    final future = await awaiter.addPending(bareJid.toString(), 'abc123', 'iq');
+    final future = await awaiter.addPending(null, 'abc123', 'iq');
 
     // Receive the wrong answer
     final stanza = XMLNode.fromString('<iq id="abc123" type="result" />');
     final result1 = await awaiter.onData(
       XMLNode.fromString('<message id="abc123" type="result" />'),
-      bareJid,
     );
     expect(result1, false);
 
     // Receive the correct answer
     final result2 = await awaiter.onData(
       stanza,
-      bareJid,
     );
     expect(result2, true);
+    expect(await future, stanza);
+  });
+
+  test('Sending a stanza to our bare JID', () async {
+    final awaiter = StanzaAwaiter(getBareJidCallback);
+
+    // "Send" a stanza
+    final future = await awaiter.addPending(bareJid, 'abc123', 'iq');
+
+    // Receive the response.
+    final stanza = XMLNode.fromString('<iq id="abc123" type="result" />');
+    await awaiter.onData(stanza);
+    expect(await future, stanza);
+  });
+
+  test(
+      'Sending a stanza to our bare JID and receiving stanza with a from attribute',
+      () async {
+    final awaiter = StanzaAwaiter(getBareJidCallback);
+
+    // "Send" a stanza
+    final future = await awaiter.addPending(bareJid, 'abc123', 'iq');
+
+    // Receive the response.
+    final stanza =
+        XMLNode.fromString('<iq from="$bareJid" id="abc123" type="result" />');
+    await awaiter.onData(stanza);
     expect(await future, stanza);
   });
 }